rename clusterdir to more descriptive baseapp
since the 'clusterdir' part is now in the base newapplication, and not cluster-specific.
