+++ IPython/parallel/apps/baseapp.py
